Added documentation and tests for MySQL errors
Added an explanation that the error occurs in the MySQL implementation of the BulkExtensions package. Added instructions on how to see the error in action by running the test. Added an image showing how the test should fail.

Added a description of the basic test `NoUpdate_JustWriteAndAssert` which should always succeed. Added a description of the `UpdateEntities_ChangedPropertiy_ExcludeFromUpdate` test, which should fail. Added an image showing the expected failure test result.

Added a detailed explanation of what happens in the `UpdateEntities_ChangedPropertiy_ExcludeFromUpdate` test. Added a code example showing how to use the `BulkUpdate` method.

It was mentioned that the MS SQL implementation works fine. It was pointed out that several similar options (PropertiesToExclude, PropertiesToExcludeOnUpdate, PropertiesToInclude, PropertiesToIncludeOnUpdate) did not work for MySQL. It was noted that no other option was found to influence the given update behaviour for MySQL.